+++ M2/Macaulay2/d/CMakeLists.txt
@@ -85,6 +85,8 @@ set(DLIST
   threads.dd
   python.d	# removed unless WITH_PYTHON
   interface.dd interface2.d
+  monoid.dd
+  monomial_ordering.dd
   texmacs.d
   boostmath.dd
   ffi.d		# removed unless WITH_FFI
