+++ M2/Macaulay2/d/Makefile.files.in
@@ -96,6 +96,8 @@ M2_SRCFILES += python-c.c python.d
 endif
 M2_DFILES += interface.dd interface2.d
 M2_DFILES += texmacs.d
+M2_DFILES += monoid.dd
+M2_DFILES += monomial_ordering.dd
 M2_DFILES += boostmath.dd
 ifneq (@LIBFFI@,no)
 M2_DFILES += ffi.d
